WIP: Update #850 so as to avoid breaking API changes & match new func
Replace the AutoFixableRuleViolation to store the link between the RuleViolation and its fixer
in the fixer itself.
This change makes it possible not only to remove all breaking API changes, but also it obeys
the purpose of linking the TextOperation with the RuleViolation, not the Fix with the RuleViolation.

The RuleViolation does not need to know how to fix all the context stuff (including node manipulation
and so on), but it needs to know which fixes ARE to be applied as text operations, because
it is only aware of text related issues, but it has nothing to do with AST changes themselves.