!Completely refactor RewriteEventsRecorder
This fully deprecates 871 and almost replaces it, solving some issues and simplifying
code that much.

871 implementation showed cases where illegal events were false reported.
For example, suppose a node having 3 children.
I one executed:
```
node.removeChild(0); // This removes child 1
node.removeChild(0); // This removes child 2, whose index had been updated because of the first removeChild.
```
This would generate a runtime exeption as it would be considered as two remove events
over the same index, which was considered to be illegal.

New implementation completely removes indexed storage and decouple the logic from maintaining
parallel node's children and rewrite events array.

This has been accomplished by realizing that the only needed thing for a rewrite event
was the old and the new value involved in a rewrite event.

The only case where two rewrite events are merged is when
oldRewriteEvent.newChildNode == newRewriteEvent.oldChildNode, because it is understand that
the new rewrite event is directly affecting the old rewrite event.